Refactor  and move files to proper folders
